widbata: make icon look full when fully charged
See https://www.desmos.com/calculator/ya0ooxshsy (showing the corners of
the rectangles to be filled); a constant 12 means there's a 2px gap to
the right of the fill when at 100% charge, but the gap on the other
sides is always 1px. A constant 11 fixes that and makes the gap 1px on
all sides at full charge.